modified:   api/app/Console/Commands/AuditRetentionPurge.php modified:   api/app/Http/Controllers/Admin/SettingsController.php modified:   api/app/Http/Controllers/Avatar/AvatarController.php modified:   api/app/Http/Controllers/Evidence/EvidenceController.php modified:   api/app/Http/Controllers/Export/ExportController.php modified:   api/app/Http/Controllers/Rbac/RolesController.php modified:   api/app/Http/Controllers/Rbac/UserRolesController.php modified:   api/app/Http/Controllers/Setup/AdminController.php modified:   api/app/Http/Controllers/Setup/AdminMfaController.php modified:   api/app/Http/Controllers/Setup/BrandingController.php modified:   api/app/Http/Controllers/Setup/DbController.php modified:   api/app/Http/Controllers/Setup/FinishController.php modified:   api/app/Http/Controllers/Setup/SmtpController.php modified:   api/app/Http/Middleware/AuthRequired.php modified:   api/app/Http/Middleware/BreakGlassGuard.php modified:   api/app/Http/Middleware/MfaRequired.php modified:   api/app/Http/Middleware/RbacMiddleware.php modified:   api/app/Http/Middleware/SetupGuard.php modified:   api/app/Http/Requests/Admin/UpdateSettingsRequest.php modified:   api/app/Http/Requests/Evidence/StoreEvidenceRequest.php modified:   api/app/Http/Requests/Rbac/StoreRoleRequest.php modified:   api/app/Jobs/GenerateExport.php modified:   api/app/Models/Evidence.php modified:   api/app/Models/Role.php modified:   api/app/Models/User.php modified:   api/app/Providers/ConfigOverlayServiceProvider.php modified:   api/app/Services/Modules/ModuleManager.php modified:   api/app/Services/Settings/SettingsService.php modified:   api/psalm.xml modified:   docs/SESSION-LOG.md modified:   docs/phase-4/PHASE-4-TASK-BREAKDOWN.md modified:   web/src/routes/admin/Audit.tsx

+++ api/app/Http/Controllers/Setup/DbController.php
@@ -22,12 +22,13 @@ final class DbController extends Controller
             return response()->json(['ok' => false, 'code' => 'SETUP_STEP_DISABLED'], 400);
         }
 
+        /** @var array{driver:string,host:string,port?:int,database:string,charset?:string,username:string,password?:string} $cfg */
         $cfg = $request->validated();
         $dsn = sprintf(
             '%s:host=%s;port=%d;dbname=%s;charset=%s',
             strtolower($cfg['driver']),
             $cfg['host'],
-            (int) ($cfg['port'] ?? 3306),
+            $cfg['port'] ?? 3306,
             $cfg['database'],
             strtolower($cfg['charset'] ?? 'utf8mb4')
         );
@@ -56,6 +57,7 @@ final class DbController extends Controller
             return response()->json(['ok' => false, 'code' => 'SETUP_STEP_DISABLED'], 400);
         }
 
+        /** @var array<string,mixed> $cfg */
         $cfg = $request->validated();
         try {
             $path = $writer->writeAtomic($cfg, (string) Config::get('core.setup.shared_config_path', '/opt/phpgrc/shared/config.php'));
